Add polkadot helm chart
This is the chart I was working on a few weeks ago. I then added more
functionality to the polkadot-k8s repo. Today I ported the most recent
changes to here. I also added a README to the chart.

The new things are (1) ability to pass a private network key and (2)
ability to pass the public ip/port combination.

It's possible to deploy locally on minikube and devspace for testing.
DEVSPACE.md has details.

The next step is to create a pulumi component resource that installs
this chart on a cluster.